+++ Sygic.Corona.Contracts/Requests/CreateContactsRequest.cs
@@ -17,6 +17,6 @@ namespace Sygic.Corona.Contracts.Requests
         public TimeSpan Duration { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
-        public int? Accuracy { get; set; }
+        public double? Accuracy { get; set; }
     }
 }
