+++ Sygic.Corona.Contracts/Requests/CreateProfileRequest.cs
@@ -8,6 +8,6 @@
         public string Locale { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
-        public int? Accuracy { get; set; }
+        public double? Accuracy { get; set; }
     }
 }
