+++ Sygic.Corona.Contracts/Requests/NotifyAreaExitRequest.cs
@@ -8,7 +8,7 @@ namespace Sygic.Corona.Contracts.Requests
         public string DeviceId { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public int Accuracy { get; set; }
+        public double Accuracy { get; set; }
         public int RecordTimestamp { get; set; }
     }
 }
